Only prune items that were published before the latest
The new fetching logic from 27b0d2c has a complicated bug: if a new item
appears partway down the feed (most frequently caused by an existing
item having its url changed), it will be one of the more recently
inserted items. This means that items closer to the start of the feed
which were actually inserted longer ago than it and will be pruned
first. Once those get pruned, we identify them as new items and we
insert them again and they become our new most recent items. The cycle
repeats from here.

I think the best way to address this is to prune a bit more
conservatively. Specifically, everything would be fine if we wait to
prune the newer-published-but-older-inserted items until after the
older-published-but-newer-inserted item is also pruned.

This updates the prune query to do that! We will now not prune any items
that are newer published than the items we're keeping. This may result
in us keeping more than 10 read rows, but it seems worth it.

(It does fail against a pathological case where the top item on the feed
is always older than any other item on a feed, but... I'll deal with
that when I encounter such a feed.)